add counter_state.dart
